Componentize the textbook preview widget
