+++ content/education/textbook-toc.md
@@ -1,0 +1,32 @@
+## Table of Contents:
+
+### Chapter 0. Introduction and Prerequisites
+- Hello World
+- Python and Jupyter Notebooks
+- Qiskit
+
+### Chapter 1. Quantum States and Qubits
+- Introduction
+- The Atoms of Computation
+- The Unique Properties of Qubits
+- Writing Down Qubit States
+- Pauli Matrices and the Bloch Sphere
+- States for Many Qubits
+
+### Chapter 2. Single-Qubit and Multi-Qubit Gates
+- Introduction
+- Quantum Gates
+- Fun with Matrices
+- The Stardard Gate Set
+- Proving Universality
+- Basic Circuit Identities
+
+### Chapter 3. Quantum Algorithms
+- Quantum Teleportation
+- Deutsch-Josza Algorithm
+- Bernstain-Vazirani Algorithm
+- Simon's Algorithm
+- Quantum Fourier Transform
+- Quantum Phase Estimation
+- Grover's Algorithm
+- Shor's Algorithm
