+++ layouts/education.vue
@@ -1,636 +1,3 @@
 <template>
-  <!-- tabindex is needed to allow hiding the menu in iOS Safari -->
-  <div class="content-root" tabindex="-1">
-    <header id="navigation">
-      <QiskitOrgMenu />
-    </header>
-    <main>
-      <header>
-        <GatesHeader
-          class="presentation"
-          main-title="Qiskit for Educators"
-          extra-position="start"
-        >
-          <p>Qiskit makes it easy to start learning quantum software to run on real quantum hardware. Teach your students with the same tools used by scientists and engineers worldwide to accelerate research towards practical applications for quantum computing.</p>
-          <template #extra>
-            <iframe
-              class="header-video"
-              width="560"
-              height="315"
-              src="https://www.youtube.com/embed/NHTDqdGfzcc"
-              frameborder="0"
-              allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-              allowfullscreen
-            />
-          </template>
-        </GatesHeader>
-      </header>
-      <div>
-        <InnerNavigation
-          class="inner-navigation"
-          :sections="[
-            { anchor: 'video-series', label: 'Coding With Qiskit Series' },
-            { anchor: 'textbook', label: 'Qiskit Textbook' },
-            { anchor: 'host-an-event', label: 'Host Qiskit Events' }
-          ]"
-        />
-        <PageSection id="video-series">
-          <h2>Coding With Qiskit Video Series</h2>
-          <p>
-            Accompany Abraham Asfaw through a series of video tutorials
-            in our YouTube Channel explaining quantum computing through
-            the use of Qiskit.
-          </p>
-          <ul class="actions">
-            <li>
-              <a class="button" href="https://www.youtube.com/playlist?list=PLOFEBzvs-Vvp2xg9-POLJhQwtVktlYGbY">
-                View all episodes
-              </a>
-            </li>
-          </ul>
-          <template #extra>
-            <iframe
-              class="episode"
-              width="560"
-              height="315"
-              src="https://www.youtube.com/embed/RrUTwq5jKM4"
-              frameborder="0"
-              allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-              allowfullscreen
-            />
-          </template>
-        </PageSection>
-        <PageSection id="textbook" extra-position="start">
-          <h2>Qiskit Textbook</h2>
-          <p>
-            Leverage the power of quantum computing using Qiskit with this
-            university course supplement covering introductory materials,
-            advanced algorithms and hardware. Include problem sets and
-            exercises for students.
-          </p>
-          <ul class="actions">
-            <li>
-              <a class="button" href="/textbook/">
-                Discover more
-              </a>
-            </li>
-          </ul>
-
-          <template #extra>
-            <section id="textbook-features">
-              <div
-                class="selector is-active"
-                data-to="toc"
-                @click.capture="activate"
-              >
-                <img src="/images/education/iconTab.svg">
-                <p>Table of Contents</p>
-              </div>
-              <div
-                class="selector"
-                data-to="live-code"
-                @click.capture="activate"
-              >
-                <img src="/images/education/iconTab.svg">
-                <p>Code Samples</p>
-              </div>
-              <section id="live-code">
-                <div class="content">
-                  <iframe
-                    height="600px"
-                    width="100%"
-                    src="https://repl.it/@delapuente/BellState?lite=true"
-                    scrolling="no"
-                    frameborder="no"
-                    allowtransparency="true"
-                    allowfullscreen="true"
-                    sandbox="allow-forms allow-pointer-lock allow-popups allow-same-origin allow-scripts allow-modals"
-                  />
-                </div>
-              </section>
-              <section id="toc" class="is-active">
-                <img class="bookmark" src="/images/education/iconBookmark.svg">
-                <div class="content">
-                  <h2>Table of Contents:</h2>
-                  <h3>Chapter 0. Introduction and Prerequisites</h3>
-                  <ul>
-                    <li>Hello World</li>
-                    <li>Python and Jupyter Notebooks</li>
-                    <li>Qiskit</li>
-                  </ul>
-                  <h3>Chapter 1. Quantum States and Qubits</h3>
-                  <ul>
-                    <li>Introduction</li>
-                    <li>The Atoms of Computation</li>
-                    <li>The Unique Properties of Qubits</li>
-                    <li>Writing Down Qubit States</li>
-                    <li>Pauli Matrices and the Bloch Sphere</li>
-                    <li>States for Many Qubits</li>
-                  </ul>
-                  <h3>Chapter 2. Single-Qubit and Multi-Qubit Gates</h3>
-                  <ul>
-                    <li>Introduction</li>
-                    <li>Quantum Gates</li>
-                    <li>Fun with Matrices</li>
-                    <li>The Stardard Gate Set</li>
-                    <li>Proving Universality</li>
-                    <li>Basic Circuit Identities</li>
-                  </ul>
-                  <h3>Chapter 3. Quantum Algorithms</h3>
-                  <ul>
-                    <li>Quantum Teleportation</li>
-                    <li>Deutsch-Josza Algorithm</li>
-                    <li>Bernstain-Vazirani Algorithm</li>
-                    <li>Simon's Algorithm</li>
-                    <li>Quantum Fourier Transform</li>
-                    <li>Quantum Phase Estimation</li>
-                    <li>Grover's Algorithm</li>
-                    <li>Shor's Algorithm</li>
-                  </ul>
-                </div>
-              </section>
-            </section>
-          </template>
-        </PageSection>
-        <PageSection id="host-an-event">
-          <h2>Host Qiskit Events</h2>
-          <p>
-            Bring Qiskit experts to your campus for guest lectures, hackathons, and other events. Guest lecture topics can range from quantum basics to advanced algorithms.
-          </p>
-          <ul class="actions">
-            <li>
-              <a class="button" href="https://www.youtube.com/playlist?list=PLOFEBzvs-Vvp2xg9-POLJhQwtVktlYGbY">
-                Request an event
-              </a>
-            </li>
-          </ul>
-          <template #extra>
-            <iframe
-              class="episode"
-              width="560"
-              height="315"
-              src="https://www.youtube.com/embed/RrUTwq5jKM4"
-              frameborder="0"
-              allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-              allowfullscreen
-            />
-          </template>
-        </PageSection>
-        <FeaturesSection id="highlights">
-          <section class="feature">
-            <img src="/images/education/iconEventsBase@3x.png">
-            <h2>Attend Exclusive Events</h2>
-          </section>
-          <section class="feature">
-            <img src="/images/education/iconCommunityBase@3x.png">
-            <h2>Join a Global Community</h2>
-          </section>
-          <section class="feature">
-            <img src="/images/education/iconDemoBase@3x.png">
-            <h2>Teach with Live Demos</h2>
-          </section>
-        </FeaturesSection>
-      </div>
-    </main>
-    <PageFooter />
-  </div>
+  <nuxt />
 </template>
-
-<script lang="ts">
-import Vue from 'vue'
-import { Component } from 'vue-property-decorator'
-import QiskitOrgMenu from '~/components/menus/QiskitOrgMenu.vue'
-import InnerNavigation from '~/components/menus/InnerNavigation.vue'
-import GatesHeader from '~/components/headers/GatesHeader.vue'
-import PageSection from '~/components/sections/PageSection.vue'
-import FeaturesSection from '~/components/sections/FeaturesSection.vue'
-import PageFooter from '~/components/footers/PageFooter.vue'
-
-@Component({
-  components: {
-    QiskitOrgMenu,
-    InnerNavigation,
-    GatesHeader,
-    PageSection,
-    FeaturesSection,
-    PageFooter
-  },
-
-  head() {
-    return {
-      title: 'Qiskit for Educators'
-    }
-  }
-})
-export default class extends Vue {
-  activate(evt: PointerEvent) {
-    const thisSelector = evt.currentTarget && evt.currentTarget as HTMLElement
-    if (!thisSelector || thisSelector.classList.contains('is-active')) {
-      return
-    }
-
-    // Change selector enabled
-    const previouslyEnabled = document.querySelector('.selector.is-active')
-    if (previouslyEnabled) {
-      previouslyEnabled.classList.remove('is-active')
-    }
-    thisSelector.classList.add('is-active')
-
-    // Dismiss the current active section
-    const self = this
-    function clearOut(evt: Event) {
-      if (evt.currentTarget) {
-        evt.currentTarget.removeEventListener('transitionend', clearOut)
-        self.clearIsOut(evt.currentTarget as HTMLElement)
-      }
-    }
-    const activeSection = document.querySelector('section.is-active')
-    if (activeSection) {
-      activeSection.addEventListener('transitionend', clearOut)
-      activeSection.classList.add('is-out')
-      activeSection.classList.remove('is-active')
-    }
-
-    // Activate the new one
-    const newSection = document.querySelector(`#${thisSelector.dataset.to}`)
-    if (newSection) {
-      newSection.classList.add('is-active')
-    }
-  }
-
-  clearIsOut(target: HTMLElement) {
-    target.classList.remove('is-out')
-  }
-}
-</script>
-
-<style lang="scss">
-@import '~/assets/scss/mixins.scss';
-@import url(~/static/css/fonts.css);
-
-:root {
-  --ibm-blue: #0a1d8f;
-  --primary-color: #242a2e;
-  --primary-color-dark: #161f25;
-  --primary-color-darkmost: #0d1a22;
-  --primary-color-light: #424c53;
-  --primary-color-lightmost: #5d6870;
-  --secondary-color: #893ffc;
-  --secondary-color-light: #a167fc;
-  --secondary-color-lightmost: #bc93fc;
-  --secondary-color-dark: #6f16fa;
-  --secondary-color-darkmost: #6105f2;
-  --body-color-light: #e0e0e0;
-  --body-color-dark: #333333;
-}
-
-.inner-navigation {
-  position: sticky;
-  top: 0;
-  z-index: 100;
-}
-
-.presentation .extra-container {
-  margin-right: 1rem;
-}
-
-.presentation .copy-container {
-  max-width: 40%;
-}
-
-.header-video {
-  width: 100%;
-  max-width: 560px;
-  height: 315px;
-  box-shadow: 0 13px 27px -5px rgba(50,50,93,.25),
-              0 8px 16px -8px rgba(0,0,0,.5),
-              0 -6px 16px -6px rgba(0,0,0,.025);
-}
-
-* {
-  padding: 0;
-  margin: 0;
-  -webkit-overflow-scrolling: touch;
-}
-
-.content-root * {
-  box-sizing: border-box;
-}
-
-html {
-  background-color: var(--primary-color);
-}
-
-html,
-body,
-#__nuxt,
-#__layout {
-  height: 100%;
-  min-height: 100%;
-}
-
-h1 {
-  font-size: 2.5rem;
-}
-
-h2 {
-  font-size: 2rem;
-}
-
-p {
-  margin-top: 1.5rem;
-}
-
-.content-root {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  min-height: 100%;
-  overflow-x: hidden;
-}
-
-.content-root > main {
-  flex-grow: 1;
-}
-
-.content-wrapper {
-  width: 100%;
-  max-width: 1100px;
-  margin: 0 auto;
-  padding: 0 2rem;
-}
-
-.button {
-  padding: 0.66rem 1rem;
-  background-color: var(--primary-color);
-  border: 2px solid var(--secondary-color);
-  font-size: 0.75em;
-  font-weight: 100;
-  color: white;
-  text-transform: uppercase;
-  text-decoration: none;
-  box-shadow: 0 4px 6px rgba(50,50,93,.11),0 1px 3px rgba(0,0,0,.08);
-  white-space: nowrap;
-  line-height: 4rem;
-}
-
-main {
-  color: var(--body-color-light);
-  background-color: var(--primary-color-darkmost);
-  background-image: linear-gradient(150deg, var(--primary-color-darkmost) 15%,var(--primary-color-dark) 70%,var(--primary-color) 94%);
-}
-
-#video-series {
-  .page-section {
-    @include framed();
-  }
-}
-
-.episode {
-  margin-left: 2rem;
-  transform: perspective(1200px) rotateY(-20deg) rotateX(5deg);
-}
-
-.episode iframe {
-  border-radius: 10px;
-  box-shadow: 0 10px 30px 0 #000000b0;
-}
-
-#textbook {
-  color: var(--body-color-dark);
-  background-color: white;
-  min-height: 600px;
-
-  .page-section {
-    @include framed();
-  }
-
-  .copy-container {
-    width: 50%;
-    background-color: white;
-    padding: 4rem 0 4rem 2rem;
-  }
-
-  .extra-container {
-    width: 60%;
-    margin-left: -10%;
-  }
-}
-
-#textbook-features {
-  position: relative;
-  min-height: calc(600px + 8rem);
-  padding: 4rem 0;
-}
-
-#textbook-features > section {
-  position: absolute;
-  width: 100%;
-}
-
-#textbook-features::before {
-  content: "";
-  display: block;
-  position: absolute;
-  bottom: 4rem;
-  width: 100%;
-  height: 100px;
-  box-shadow: 0 23px 35px 0 rgba(10, 0, 50, 0.35);
-  border-radius: 8px;
-  transform-origin: center bottom;
-  transform: scale(0.95);
-  z-index: 1;
-}
-
-#textbook-features > section {
-  opacity: 0;
-  position: absolute;
-  bottom: 5rem;
-  height: 600px;
-  background: white;
-  border-radius: 8px;
-  padding: 2rem;
-  transform-origin: center bottom;
-  transform: scale(0.95) translateY(1rem);
-  transition: transform 300ms, opacity 300ms;
-}
-
-#textbook-features > section.is-active {
-  opacity: 1;
-  z-index: 2;
-  transform: scale(1);
-  box-shadow: 0 23px 35px 0 rgba(10, 0, 50, 0.35);
-}
-
-#textbook-features > .is-out {
-  opacity: 0;
-  z-index: 2;
-  transform: scale(1.3) translateY(-2rem);
-  transition: transform 400ms, opacity 400ms;
-}
-
-#textbook-features .bookmark {
-  position: absolute;
-  top: -16px;
-  left: 1rem;
-  z-index: 3;
-  width: 3.8rem;
-  height: auto;
-}
-
-#textbook-features .selector {
-  position: absolute;
-  bottom: 5rem;
-  left: calc(100% - 2.8rem);
-  transition: transform 200ms;
-  cursor: pointer;
-}
-
-#textbook-features .selector img {
-  height: 4rem;
-  width: auto;
-}
-
-#textbook-features .selector p {
-  margin: 0;
-  position: absolute;
-  bottom: 0.7rem;
-  right: 1rem;
-  writing-mode: vertical-rl;
-  height: 100%;
-  color: white;
-  font-size: 0.7rem;
-  font-weight: bold;
-  transform: rotate(180deg);
-}
-
-#textbook-features .selector:nth-child(2) {
-  bottom: 9rem;
-  left: calc(100% - 2.8rem);
-}
-
-#textbook-features .selector.is-active,
-#textbook-features .selector:hover {
-  transform: translateX(1rem);
-}
-
-#textbook-features .content {
-  z-index: 1;
-  max-height: 100%;
-  overflow-y: auto;
-  padding-left: 3.5rem;
-  font-size: 0.8rem;
-}
-
-#textbook-features .content ul {
-  margin: 0.5rem 0 0 2rem;
-}
-
-#textbook-features .content h3 {
-  margin: 1rem 0 0 0;
-  color: var(--secondary-color);
-}
-
-#textbook-features #live-code,
-#textbook-features #live-code * {
-  margin: 0;
-  padding: 0;
-  overflow: hidden;
-}
-
-#host-an-event {
-  .page-section {
-    @include framed();
-  }
-}
-
-#highlights {
-  .features-section {
-    display: flex;
-    flex-direction: row;
-  }
-}
-
-#highlights .feature {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.feature {
-  box-sizing: border-box;
-  border: 4px solid var(--secondary-color);
-  border-radius: 10px;
-  margin: 0 1rem 0 0;
-  padding: 2rem 1rem;
-}
-
-.feature:last-child {
-  margin-right: 0;
-}
-
-.feature img {
-  width: 50%;
-  padding-bottom: 2rem;
-}
-
-.feature h2 {
-  font-size: 1.1rem;
-  text-align: center;
-}
-
-.feature p {
-  text-align: center;
-  margin-top: 1rem;
-}
-
-@media (max-width: 600px) {
-  #intro,
-  #video-series .content-wrapper {
-    display: block;
-  }
-
-  iframe,
-  section > aside,
-  #textbook aside {
-    display: none;
-  }
-
-  #highlights {
-    display: none;
-  }
-
-  #video-series {
-    margin-top: 0;
-  }
-
-  #textbook {
-    min-height: unset;
-  }
-
-  #intro .description,
-  #video-series .description,
-  #textbook .description {
-    width: 100%;
-    max-width: 100%;
-    padding-left: 0;
-  }
-
-  footer {
-    margin-top: 0;
-  }
-
-  footer .content-wrapper {
-    display: block;
-  }
-
-  footer h2 {
-    margin: 4rem 0 0;
-  }
-
-  footer .content-wrapper > section:last-child {
-    margin-right: 2rem;
-  }
-}
-</style>
